+++ frontend/src/Components/Landing/TestomonialSection.tsx
@@ -5,45 +5,7 @@ import { useRef } from 'react'
 import BlogSection from './Blogsection'
 
 import Image from 'next/image'
-
-const Testimonials = [
-	{
-		name: 'Nguyễn Văn A',
-		avatar: '/landing/1.jpg',
-		content:
-			'Ứng dụng Health Care rất tiện lợi, giúp tôi theo dõi sức khỏe hàng ngày và đặt lịch khám dễ dàng.',
-	},
-	{
-		name: 'Trần Thị B',
-		avatar: '/landing/2.jpg',
-		content:
-			'Tôi cảm thấy an tâm hơn khi sử dụng Health Care để theo dõi chu kỳ và lịch tiêm phòng cho con.',
-	},
-	{
-		name: 'Lê Minh Cường',
-		avatar: '/landing/3.jpg',
-		content:
-			'Dịch vụ tư vấn trực tuyến rất nhanh chóng và chuyên nghiệp. Tôi đã được bác sĩ hỗ trợ kịp thời.',
-	},
-	{
-		name: 'Phạm Quỳnh Anh',
-		avatar: '/landing/4.jpg',
-		content:
-			'Ứng dụng có giao diện dễ sử dụng, phù hợp với cả người lớn tuổi trong gia đình tôi.',
-	},
-	{
-		name: 'Hoàng Thanh Tùng',
-		avatar: '/landing/5.jpg',
-		content:
-			'Tôi đánh giá cao tính năng theo dõi lịch sử khám bệnh và nhắc nhở dùng thuốc của Health Care.',
-	},
-	{
-		name: 'Nicolai Hong',
-		avatar: '/landing/6.jpg',
-		content:
-			'Health Care giúp tôi đặt lịch tư vấn sản phụ khoa dễ dàng, không cần chờ đợi lâu tại bệnh viện.',
-	},
-]
+import Testimonials from '@/Constants/Testomonial'
 
 /**
  * Section for testimonials. Scroll-based animation for each testimonial.
@@ -104,7 +66,7 @@ export default function TestimonialsSection() {
 							>
 								<Image
 									src={item.avatar}
-									alt='consultant'
+									alt={`Portrait of ${item.name}`}
 									className='object-cover h-[400px]'
 									width={500}
 									height={500}
