+++ frontend/src/Constants/Testomonial.ts
@@ -1,0 +1,40 @@
+const Testimonials = [
+	{
+		name: 'Nguyễn Văn A',
+		avatar: '/landing/1.jpg',
+		content:
+			'Ứng dụng Health Care rất tiện lợi, giúp tôi theo dõi sức khỏe hàng ngày và đặt lịch khám dễ dàng.',
+	},
+	{
+		name: 'Trần Thị B',
+		avatar: '/landing/2.jpg',
+		content:
+			'Tôi cảm thấy an tâm hơn khi sử dụng Health Care để theo dõi chu kỳ và lịch tiêm phòng cho con.',
+	},
+	{
+		name: 'Lê Minh Cường',
+		avatar: '/landing/3.jpg',
+		content:
+			'Dịch vụ tư vấn trực tuyến rất nhanh chóng và chuyên nghiệp. Tôi đã được bác sĩ hỗ trợ kịp thời.',
+	},
+	{
+		name: 'Phạm Quỳnh Anh',
+		avatar: '/landing/4.jpg',
+		content:
+			'Ứng dụng có giao diện dễ sử dụng, phù hợp với cả người lớn tuổi trong gia đình tôi.',
+	},
+	{
+		name: 'Hoàng Thanh Tùng',
+		avatar: '/landing/5.jpg',
+		content:
+			'Tôi đánh giá cao tính năng theo dõi lịch sử khám bệnh và nhắc nhở dùng thuốc của Health Care.',
+	},
+	{
+		name: 'Nicolai Hong',
+		avatar: '/landing/6.jpg',
+		content:
+			'Health Care giúp tôi đặt lịch tư vấn sản phụ khoa dễ dàng, không cần chờ đợi lâu tại bệnh viện.',
+	},
+]
+
+export default Testimonials
